Oregon Redistricting File

This data contains geographic, demographic, and political information on the 1,071 census tracts of the state of Oregon.

Usage

data("oregon")

Format

sf object

county

The county the tract belongs to.

cd_2020

The 2210 congressional district assignment for the tract.

pop

The total population of the tract, according to the 2020 Census.

pop_white

The non-Hispanic white population of the precinct.

ndv

Average number of votes for Democratic candidates in recent statewide elections.

nrv

Average number of votes for Republican candidates in recent statewide elections.

geometry

The sf geometry column containing the geographic information.

Examples

data(oregon)
oregon[, 1:6]
#> Simple feature collection with 1071 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2163029 ymin: -185907.3 xmax: 2814746 ymax: 293445.7
#> Projected CRS: NAD83(HARN) / Oregon North
#> # A tibble: 1,071 × 7
#>    county cd_2020   pop pop_white   ndv   nrv                           geometry
#>    <chr>    <int> <int>     <int> <dbl> <dbl>                 <MULTIPOLYGON [m]>
#>  1 Baker        2  2844      2552  369. 1312. (((2705037 125011.4, 2710726 1212…
#>  2 Baker        2  3191      2738  498. 1251. (((2711347 127396.3, 2711327 1265…
#>  3 Baker        2  2745      2390  392. 1247. (((2760731 96173.67, 2759438 8633…
#>  4 Baker        2  3088      2758  148.  402. (((2711327 126595.8, 2711347 1273…
#>  5 Baker        2  2749      2380  464. 1127  (((2709192 128933.2, 2698934 1477…
#>  6 Baker        2  2051      1846  336.  932. (((2714628 150906.5, 2731447 1515…
#>  7 Benton       4  4344      3161 1436.  458  (((2279770 100406.1, 2279322 1004…
#>  8 Benton       4  3409      2308 1188.  394  (((2280644 102610.9, 2280706 1026…
#>  9 Benton       4  5256      4233 2049.  908  (((2274712 99335.66, 2274240 1013…
#> 10 Benton       4  2968      2340 1709.  373. (((2277555 106896, 2279606 105983…
#> # ℹ 1,061 more rows